requirements.txt
pillow
kurigram
tgcrypto